# docs/openapi/openAPI-3-schema.YAML
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v3.4.0
  description: SPYDERISK System Modeller (SSM) REST API definitions for domain models, user system models and usage by other applications.
  contact:
    name: University of Southampton IT Innovation Centre
    url: http://www.it-innovation.soton.ac.uk/contact-us
    email: info@spyderisk.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8081/system-modeller
    description: Local server
paths:
...
  /models/{modelId}/relations/{relationId}:
    get:
      tags:
      - relation-controller
      description: Get information about a relation.
      operationId: getRelation
      parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: the relation contained in the model
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Relation'
    put:
      tags:
      - relation-controller
      description: Update information about a relation.
      operationId: updateRelation
      parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Relation'
        required: true
      responses:
        "200":
          description: the updated relation contained in the model
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Relation'
    delete:
      tags:
      - relation-controller
      description: Delete a relation given the model and relation IDs.
      operationId: deleteRelationInModel
      parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: a JSON object describing the RelationUi with the given relation
            id
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DeleteRelationResponse'
...
  /models/{modelId}/relations:
    get:
      tags:
      - relation-controller
      description: Get a list of Relation present in the model.
      operationId: listModelRelations
      parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: a list of relations contained in the model
          content:
            '*/*':
              schema:
                uniqueItems: true
                type: array
                items:
                  $ref: '#/components/schemas/Relation'
    post:
      tags:
      - relation-controller
      description: Submit a new relation to be created.
      operationId: createRelation
      parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Relation'
        required: true
      responses:
        "200":
          description: "the persisted relation, error otherwise"
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateRelationResponse'
...
  /administration/users/{userId}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    AssetDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        type:
          type: string
        asserted:
          type: boolean
        visible:
          type: boolean
        iconX:
          type: integer
          format: int32
        iconY:
          type: integer
          format: int32
        minCardinality:
          type: integer
          format: int32
        maxCardinality:
          type: integer
          format: int32
        population:
          type: string
        controlSets:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ControlSet'
        misbehaviourSets:
          uniqueItems: true
          type: array
          items:
            type: string
        inferredAssets:
          uniqueItems: true
          type: array
          items:
            type: string
        trustworthinessAttributeSets:
          uniqueItems: true
          type: array
          items:
            type: string
        id:
          type: string
    AssetGroupDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        left:
          type: string
        top:
          type: string
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        expanded:
          type: boolean
        assetIds:
          type: array
          items:
            type: object
        name:
          type: string
        id:
          type: string
    ComplianceSetDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        systemThreats:
          uniqueItems: true
          type: array
          items:
            type: string
        compliant:
          type: boolean
        id:
          type: string
    ComplianceThreatDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        pattern:
          $ref: '#/components/schemas/Pattern'
        threatensAssets:
          type: string
        type:
          type: string
        resolved:
          type: boolean
        acceptanceJustification:
          type: string
        controlStrategies:
          type: object
          additionalProperties:
            type: string
            enum:
            - BLOCK
            - MITIGATE
            - TRIGGER
        id:
          type: string
    ControlSet:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        control:
          type: string
        assetUri:
          type: string
        assetId:
          type: string
        proposed:
          type: boolean
        assertable:
          type: boolean
        workInProgress:
          type: boolean
        coverageLevel:
          type: string
        coverageAsserted:
          type: boolean
        id:
          type: string
    ControlStrategyDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        blockingEffect:
          $ref: '#/components/schemas/Level'
        enabled:
          type: boolean
        mandatoryControlSets:
          uniqueItems: true
          type: array
          items:
            type: string
        optionalControlSets:
          uniqueItems: true
          type: array
          items:
            type: string
        threatCsgTypes:
          type: object
          additionalProperties:
            type: string
            enum:
            - BLOCK
            - MITIGATE
            - TRIGGER
        id:
          type: string
    Level:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        value:
          type: integer
          format: int32
        id:
          type: string
    Link:
      type: object
      properties:
        fromAsset:
          type: string
        fromAssetLabel:
          type: string
        fromRole:
          type: string
        fromRoleLabel:
          type: string
        toAsset:
          type: string
        toAssetLabel:
          type: string
        toRole:
          type: string
        toRoleLabel:
          type: string
        type:
          type: string
        typeLabel:
          type: string
    MisbehaviourSet:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        misbehaviour:
          type: string
        misbehaviourLabel:
          type: string
        asset:
          type: string
        assetLabel:
          type: string
        visible:
          type: boolean
        impactLevel:
          $ref: '#/components/schemas/Level'
        likelihood:
          $ref: '#/components/schemas/Level'
        riskLevel:
          $ref: '#/components/schemas/Level'
        impactLevelAsserted:
          type: boolean
        directCauses:
          uniqueItems: true
          type: array
          items:
            type: string
        indirectCauses:
          uniqueItems: true
          type: array
          items:
            type: string
        rootCauses:
          uniqueItems: true
          type: array
          items:
            type: string
        directEffects:
          uniqueItems: true
          type: array
          items:
            type: string
        id:
          type: string
    ModelDTO:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        description:
          type: string
        domainGraph:
          type: string
        domainVersion:
          type: string
        validatedDomainVersion:
          type: string
        valid:
          type: boolean
        riskLevelsValid:
          type: boolean
        calculatingRisks:
          type: boolean
        canBeEdited:
          type: boolean
        canBeShared:
          type: boolean
        risk:
          $ref: '#/components/schemas/Level'
        threats:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ThreatDTO'
        controlStrategies:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ControlStrategyDTO'
        complianceThreats:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ComplianceThreatDTO'
        complianceSets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ComplianceSetDTO'
        misbehaviourSets:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/MisbehaviourSet'
        twas:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TrustworthinessAttributeSet'
        controlSets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ControlSet'
        userId:
          type: string
        editorId:
          type: string
        created:
          type: string
          format: int64
        modified:
          type: string
          format: int64
        modifiedBy:
          type: string
        levels:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/Level'
        riskVector:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RiskLevelCount'
        assets:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AssetDTO'
        relations:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Relation'
        groups:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AssetGroupDTO'
        loadingId:
          type: string
        validating:
          type: boolean
    Node:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        asset:
          type: string
        role:
          type: string
        assetLabel:
          type: string
        roleLabel:
          type: string
        id:
          type: string
    Pattern:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        parent:
          type: string
        parentLabel:
          type: string
        nodes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Node'
        links:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Link'
        id:
          type: string
    Relation:
      type: object
      properties:
        from:
          type: string
        fromID:
          type: string
        to:
          type: string
        toID:
          type: string
        type:
          type: string
        label:
          type: string
        sourceCardinality:
          type: integer
          format: int32
        targetCardinality:
          type: integer
          format: int32
        asserted:
          type: boolean
        visible:
          type: boolean
        immutable:
          type: boolean
        hidden:
          type: boolean
        inferredAssets:
          uniqueItems: true
          type: array
          items:
            type: string
        id:
          type: string
    RiskLevelCount:
      type: object
      properties:
        level:
          $ref: '#/components/schemas/Level'
        count:
          type: integer
          format: int32
    SemanticEntity:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        id:
          type: string
    ThreatDTO:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        pattern:
          $ref: '#/components/schemas/Pattern'
        threatensAssets:
          type: string
        type:
          type: string
        resolved:
          type: boolean
        acceptanceJustification:
          type: string
        misbehaviours:
          uniqueItems: true
          type: array
          items:
            type: string
        indirectEffects:
          uniqueItems: true
          type: array
          items:
            type: string
        secondaryThreat:
          type: boolean
        secondaryEffectConditions:
          uniqueItems: true
          type: array
          items:
            type: string
        controlStrategies:
          type: object
          additionalProperties:
            type: string
            enum:
            - BLOCK
            - MITIGATE
            - TRIGGER
        entryPoints:
          uniqueItems: true
          type: array
          items:
            type: string
        likelihood:
          $ref: '#/components/schemas/Level'
        riskLevel:
          $ref: '#/components/schemas/Level'
        rootCause:
          type: boolean
        id:
          type: string
    TrustworthinessAttributeSet:
      type: object
      properties:
        uri:
          type: string
        label:
          type: string
        description:
          type: string
        parents:
          uniqueItems: true
          type: array
          items:
            type: string
        asset:
          type: string
        attribute:
          $ref: '#/components/schemas/SemanticEntity'
        causingMisbehaviourSet:
          type: string
        assertedTWLevel:
          $ref: '#/components/schemas/Level'
        inferredTWLevel:
          $ref: '#/components/schemas/Level'
        twLevelAsserted:
          type: boolean
        visible:
          type: boolean
        id:
          type: string
...
    CreateRelationResponse:
      type: object
      properties:
        relation:
          $ref: '#/components/schemas/Relation'
        model:
          $ref: '#/components/schemas/ModelDTO'
...
    DeleteRelationResponse:
      type: object
      properties:
        relations:
          uniqueItems: true
          type: array
          items:
            type: string
        valid:
          type: boolean
        empty:
          type: boolean
...